Beginnings of per-level cost modeling
The total time required for each shard construction
is now measured, and hooks have been added to
InternalLevel to support the creation and use of
per-level cost models.

The appropriate calls to these functions are now
made during reconstruction, but the scheduling
process doesn't yet call them, nor are the models
actually implemented.