Implemented dynamic insertion rate adjustment
The insertion rate will now be updated on the fly
to block inserts based on the sum of the predicted
runtime and record counts for currently active
reconstructions.

The scheduling system now also calls into the
runtime prediction system and handles this data
appropriately.

All that remains is to implement the prediction
models themselves--then we should be good to test.